game files refactored
