--- backend/chat/Dockerfile
@@ -1,15 +1,0 @@
-FROM node:23-bullseye-slim
-
-WORKDIR /app
-
-COPY package*.json ./
-
-RUN npm clean-install
-
-COPY . .
-
-RUN npx prisma generate
-
-RUN npm install -g ts-node
-
-CMD ["ts-node", "src/index.ts"]
