--- backend/chat/package.json
@@ -1,19 +1,0 @@
-{
-    "name": "chat-service",
-    "version": "1.0.0",
-    "main": "src/index.ts",
-    "scripts": {
-        "start": "node src/index.ts"
-    },
-    "devDependencies": {
-        "@types/node": "^24.0.13",
-        "prisma": "^6.11.1",
-        "ts-node": "^10.9.2",
-        "typescript": "^5.8.3"
-    },
-    "dependencies": {
-        "@fastify/type-provider-typebox": "^5.1.0",
-        "@prisma/client": "^6.11.1",
-        "fastify": "^5.4.0"
-    }
-}
