--- backend/chat/prisma/schema.prisma
@@ -1,22 +1,0 @@
-generator client {
-    provider = "prisma-client-js"
-}
-
-datasource db {
-    provider = "sqlite"
-    url      = "file:./prisma/chat-db.sqlite"
-}
-
-model Message {
-    id        Int      @id @default(autoincrement())
-    content   String
-    userId    Int
-    roomId    String
-    createdAt DateTime @default(now())
-}
-
-model Room {
-    id        String   @id
-    name      String
-    createdAt DateTime @default(now())
-}
